do not include directory when moh is custom
in asterisk/musiconhold.conf, the directory conflicts
with the application. Solution: don't add the directory

However, because of technical limitations in staticconf,
this makes it hard to manage when editing a category and changing
the mode. To camouflage the problem, a mode is no longer changeable
when editing a category.