don't delete func keys of other related users
problem: when deleting a user with n func keys,
delete_func_key_for_dst_user also deleted func key destinations
for the other users in the list, which is a BAD THING.
All we want is to dissociate the func_key in func_key_mapping.
This commit removes the faulty code that deletes the other
func keys.