add upgrade action on plugin package
Right now, is for example you have a plugin "xivo-yealink-v80" with a
firmware "T41P-fw", then if the installed version of the firmware is not
the same as the installable version, then an upgrade icon is shown next
to the uninstall icon for the "T41P-fw" firmware.

Work in progress.