cel ws: order by id instead of eventtime
This has a large performance impact (a positive one that is). This is
especially true when you have lots of CEL in the table and are
requesting an ID in the not close to the max ID nor the min ID. In this
case you can get a 100x increase in performance. For the case where you
are requesting an ID near the maximum ID, this doesn't change much.

This does changes the interface of the web service a little bit, i.e.
the CEL are not ordered by eventtime anymore but by ID. The order is
usually the same, but not in all cases. That said, there was some edge
cases with the previous behavior that make it a bit cumbersome to use
robustly.