directory display: respect the column order
display columns are stored in a map with keys being integers incremented
by 10.

example:

10: [Name, name, '', name],
20: [Number, number, '', number],
...
70: [Location, '', '', location],

this map is stored in a json string in the database. If for some reason
the keys are not sorted in the dictionnary representation as string,
which is equivalent since dict are not ordered, the web interface ignore
the column indexes and breaks the original display when saving a
subsequent modification, overwritting the old (good) indexes with the
new ones.