dwho_gat: update reference related code
On XiVO 15.18, dwho_gat::get used to return a references to an array
stored in dwho_gatearray. This was not working anymore, because:

* code was updated incorrectly in 58064b40e438cce6d1b6e6ea0a4f3ca6f430f9dd
* doing a "return(($ref = &$r))" never returns a reference to $r in php 5.6

So update the code, and update the return format to use the
without-parenthese form (which is mandatory for the code to work
correctly).

Also update the return format everywhere else in the file, so that it's
similar.