jessie: update jquery-ui usage
The problem was: function keys of type "custom" could not be saved
anymore. That's because just including jquery.ui.autocomplete.js is not
working anymore with jquery-ui 1.10.

This commit also replace the "ui-lightness" theme to the "smoothness"
theme, since the ui-lightness theme on XiVO is for jquery-ui 1.8. Theme
would need to be updated.

Also, clean a bit the multiple inclusion; jquery.js, jquery-ui.js and
jquery-ui.css are all included from meta.php.