remove deprecated preg_replace() /e modifier
The preg_replace /e modifier has been deprecated in PHP 5.5, and is
removed in PHP 7. Instead of updating the code, just remove it, since it
was not useful.

For example, when it was receiving a header:

   location: foo

it would transfer it into:

   Location: foo.

But anyway, this is just used to interact with xivo-provd, and
only to fetch the Location header, and the Location header is always
written as "Location" in xivo-provd.