use summary view to fetch information on users
when requesting a list of all users, the user app
fetches the main line for every user. This means
that for every user, 5 requests were sent to confd. This
translated into a heavy performance hit.

The information that the webi needs to display is now all regrouped
into a 'summary' view, meaning that only one request to confd is needed
for fetching information on all users