line: remove broken "where protocol" in dso
A "where protocol" filter could be pass to the dso only via the line web
service (at least in recent past), but the dso was not handling the
filter correctly: it was checking if a key *protocols* existed, but then
fetched the value of the *procotol* key, which was of course not
defined.

So remove all the code related to this.