line: remove more broken code
In the lines listing page, the leftmost icon showed if the telephone was
disabled, "enabled and registered" or "enabled and not-registered", but
the last case was relying asterisk realtime being enabled, which is not
possible in xivo anymore since XiVO 1.2. So remove the code and use the
standard icons for enable/disable.