line: remove some unused code
And clarify what was really being done: since the users were dissociated
from the line *before* listing the extensions associated to the line,
the extensions associated to an incall *were not* listed in the
association, and so testing if the associated extension was in an
internal context was always true.

This keep the behaviour of when deleting a line, the incall is *not
deleted*. Note that dissociating the line from the extension via confd
when the extension is associated to an incall deletes the incall.

I feel like xivo-confd *should not* list the extensions associated to
incalls as extensions associated to the line. When that will be the
case, the code won't need to be changed, and the users could be
dissociated before the extensions and the behaviour would be the same.