line: remove code related to SIP "hidden options"
Was not useful. For the SIP "call-limit" option, the idea was to hide it
since it was mostly used for the "simultaneous" calls of the associated
user, but the administrator could still change the value anyway, he was
just not seeing the current value.

For SCCP, this is still necessary, since we don't display the options at
all in the form yet.