simplify regex
There was a small mistake in the regex, i.e. the character class should
have been something like "[a-zA-Z]", and probably should have use a "+"
instead of "*" to be on the safe side, but instead replace that with a
plain "Asterisk" pattern since this is stable and this is really the
thing we are looking for.