fix erronous condition
chk_active should return true if the last user activity happened less
than 60 seconds ago, not if it happened more than 60 seconds ago.

By fixing this bug, xivo_authent_xivo needed to be modified because it
had actually a dependency on this bug. To fix the bug, store the
token in _SESSION['token'] instead of _SESSION['_USR']->token.

If we really want to store the token in _SESSION['_USR'], then
xivo_user_abstract need to be modified in the load_by_xxx or _set_info
method, since it's the _set_info method that is responsible to set
_SESSION['_USR'] and xivo_authent_xivo should definitely not set it.