outcall: always order trunks by priority
The web interface was actually not forcing the database to return the
outcall of a trunk in the priority order and was not ordering them
explicitly by priority: it was in fact relying on an undefined behavior
which was actually working must of the time. But this changed a bit now
that xivo-confd can create outcalls, so define the order explicitly.