paging: delete paging members before paging itself
This has become necessary with the addition of a foreign key between the
paginguser table and userfeatures table.

Also, fix the reverse delete which wasn't working (because the origin
array contained false values, because the origin_list was not set
properly on the module, because it was using a custom dso method) now
that the order makes it more probable to need a working implementation
of reverse_delete.