1364 - display Polaris in footer and configuration page
Instead of displaying
 XIVO - Version : 2017.LTS2 (2017.11.00) Polaris
which was a concatenation of
 XIVO_SOFT_LABEL	    XIVO
 $this->bbf('info_version') Version
 XIVO_SOFT_VERSION	    2017.LTS2 (2017.11.00)
 XIVO_SOFT_CODENAME	    Polaris
we decided to display only :
 XIVO - Polaris (2017.11.00)

Polaris and 2017.11.00 are defined in xivo package

Translation for info_version was not used anymore so it was removed.
XIVO_SOFT_CODENAME also.

It will also ease when releasing Intermediate Version (only xivo package
will change, no longer xivo-web-interface).

It includes change for Configuration page too.